Breadth-first Search and Backtracking
There are two stages for solving this problem.In first stage, we can
use BFS to generate inverse adjacency list.And then, in second stage,
it is necessary use DFS(aka, Backtracking) to generate all shortest
paths. Futhermore,the method in second stage is suit for the above
data-structure which had generated in first stage.